Increasing readibility
Setting lists to just 1 space between list item and it's contents. 
Changing headers to wikiHow "Step 1" to make it clearer how the user is supposed to go through the document.
Beginning to change the shell commands to single line copy and paste events to make it easier for the user. I'd be half tempted to write a small bash script to do this automatically, but it's good practice I guess!

In 'Core Packages"
Changed -S to -Syu as recommended by ArchLinux, fully understand that -S should drag in necessary packages but -Syu is preferred in case of other changes.

Stopped at Step3